requirements.txt
flask-restx==0.2.0
Flask-Script==2.0.6
Flask==1.1.2
gunicorn==20.0.4
python-dotenv==0.14.0
requests==2.24.0
